Moved operations around to be more consistent and potentially offer type checking (if PyCharm worked properly).

+++ haikuwriters/scoring/tree.py
@@ -60,7 +60,7 @@ Empty = ScoreTree(None)
 
 
 class Score(ScoreTree):
-    def __init__(self, score:int):
+    def __init__(self, score:float):
         self._score = score
         super().__init__(score)
